perimeter_of_circle

+++ 04.Input().py/12.perimeter_of_circle.py
@@ -1,4 +1,4 @@
-r = int(input("enter the radious of circle:"))
+r = float(input("enter the radious of circle:"))
 perimeter =(2*22/7 * r)
 
 print("perimeter of circle is:",perimeter)
